task: project restructure
