+++ src/GridForge/Grids/Managers/GlobalGridManager.cs
