+++ src/GridForge/Grids/Managers/GridObstacleManager.cs
