+++ src/GridForge/Grids/Managers/GridOccupantManager.cs
