+++ src/GridForge/Grids/Nodes/ScanCell.cs
